Implement the UBIFS flashing type
Implement the flashing procedure for UBIFS partitions.

The difference is on the command used to flash the partition:
ubiupdatevol command performs an erase so it is not necessary to
erase the partition before flashing it. It also requires the size
of the file to use a file as input.

To be able to know if the partition the user wants to flash is
a MTD or an UBI partition, create a "is_mtd" field in the fwupgrade
structure.

Signed-off-by: Mylène Josserand <mylene.josserand@free-electrons.com>
[Thomas:
 - Switch the is_mtd boolean to a enum { TYPE_MTD, TYPE_UBI } type field
   in 'struct fwupgrade_action'.
 - Rename 'nandwrite_pipe' to 'flash_pipe', since nandwrite is only used
   for MTD flashing, not UBI flashing.
 - Use strcmp instead of strstr to test the partition type when parsing
   the configuration file.]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>